# backend/src/routes/leave-requests.js
const express = require('express');
const { authenticateToken } = require('../middleware/auth');

const router = express.Router();

// Apply authentication to all routes
router.use(authenticateToken);

// Placeholder routes - will be implemented based on LeaveRequest model
router.get('/', (req, res) => {
  res.status(200).json({
    data: [],
    message: 'Leave requests endpoint - coming soon',
  });
});

router.post('/', (req, res) => {
  res.status(201).json({
    message: 'Create leave request endpoint - coming soon',
  });
});

router.get('/:leaveId', (req, res) => {
  res.status(200).json({
    message: 'Get leave request endpoint - coming soon',
  });
});

router.post('/:leaveId/approve', (req, res) => {
  res.status(200).json({
    message: 'Approve leave request endpoint - coming soon',
  });
});

router.post('/:leaveId/reject', (req, res) => {
  res.status(200).json({
    message: 'Reject leave request endpoint - coming soon',
  });
});

module.exports = router;